+++ src/pages/AuthPage.tsx
@@ -1,0 +1,3 @@
+export const AuthPage = () => {
+  return <h1>Auth Page</h1>
+}
