+++ src/pages/NotFoundPage.tsx
@@ -1,0 +1,3 @@
+export const NotFoundPage = () => {
+	return <h1>404 - Page Not Found</h1>
+}
